Final status: Liquibase xml needs to be updated.

+++ src/main/java/am/davsoft/sfl_assessment/repository/ProductRepository.java
@@ -1,11 +1,11 @@
 package am.davsoft.sfl_assessment.repository;
 
-import am.davsoft.sfl_assessment.entity.Product;
+import am.davsoft.sfl_assessment.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 /**
  * @author David Shahbazyan
- * @since Jul 18, 2020
+ * @since Jul 20, 2020
  */
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 }
